LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-remove-expected.txt


FAIL First meta applies. assert_equals: expected "dark" but got "light"
FAIL Second meta applies after first one is removed. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "auto"
FAIL Initial color-scheme with both meta elements removed. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "auto"
